refactor: move middleware.tsx to middleware/index.ts
